Add really basic scale factor
I stole this method from SimpleTweaks. Grabs the scale factor that's set and applies it to some of the UI so XIVCombo no longer cuts off the save/save and close buttons if people have a large scale factor.

Also adjusted two of the hint paths from assumed locations to %appdata%